Ignore case when globbing file names
I was unknowingly relying on the fact that nvim defaulted to ignoring file name case
if you are on an OS with a case insensitive file system.

Now I'm explicitly setting it before search for matching files and then
putting it back to its original setting.

Interestingly, the sort order is a combined a-z if you use
fileignorecase but if you use wildignorecase it sorts it A-Z followed by
a-z