Fix firing before other plugs are ready
Playing around with using `folke/snacks` instead of
`stevearc/dressing.nvim` for `vim.ui.select` and noticed `actually`
fired before `snacks` no matter how I tried to order them or define
dependencies.

For now I'm scheduling the select to give time for other plugins to be
ready.

Also escaping the filename in two places to be consistent, instead of
just one and not the other.